refactor: moved ogf, skl plugins
