+++ io_scene_xray/ogf/exp.py
@@ -4,13 +4,13 @@ from bmesh.ops import triangulate
 import bpy
 import mathutils
 
-from .xray_io import ChunkedWriter, PackedWriter
-from .fmt_ogf import Chunks, ModelType, VertexFormat
-from .utils import is_exportable_bone, find_bone_exportable_parent, AppError, \
+from ..xray_io import ChunkedWriter, PackedWriter
+from .fmt import Chunks, ModelType, VertexFormat
+from ..utils import is_exportable_bone, find_bone_exportable_parent, AppError, \
     fix_ensure_lookup_table, convert_object_to_space_bmesh, \
     calculate_mesh_bbox, gen_texture_name
-from .utils import is_helper_object
-from .xray_motions import MATRIX_BONE_INVERTED
+from ..utils import is_helper_object
+from ..xray_motions import MATRIX_BONE_INVERTED
 
 
 def calculate_mesh_bsphere(bbox, vertices):
